// src/app/modules/Recipes/recipes.interface.ts
export type IRecipes = {
  id?: number;
  name: string;
  category_id?: number | null;
  yield_quantity?: string | null;
  yield_unit?: string | null;
  total_weight?: string | null;
  tag_ids?: number[] | null;
  ingredients?: any[] | null;
  description?: string | null;
  note?: string | null;
  created_at?: Date;
  updated_at?: Date;
  category?: any | null;
  tags?: any[] | null;
  images?: string[];
  recipe_code?: string;
  instruction?: string;
  estimated_time?: string;
  price?: number;
};